simple_todo/requirements.txt
django==3.2.6

djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
django-debug-toolbar==3.2
drf-yasg==1.20.0

psycopg2-binary==2.8.6

#utils
factory-boy==3.2.0
